LayoutTests/imported/w3c/web-platform-tests/css/css-multicol/multicol-gap-animation-001-expected.txt


FAIL column-gap property is animatable assert_equals: expected "100px" but got "0px"
